+++ mainlib/maingui/settingsdialog.h
@@ -73,6 +73,14 @@ private slots:
 
     void on_videoShowWholeVideoOption_toggled(bool checked);
 
+    void on_maximumInclineSpinBox_valueChanged(int maximumDownhillPercentage);
+
+    void on_maximumDeclineSpinBox_valueChanged(int maximumDownhillPercentage);
+
+    void on_powerForElevationCorrectionSpinBox_valueChanged(int powerForElevationCorrection);
+
+    void on_difficultySettingSlider_valueChanged(int value);
+
 private:
     void reset();
     void fillSensorSettingsComboBox();
@@ -85,6 +93,8 @@ private:
     void fillVideoDisplayOptions();
     void fillWeights();
     void fillPowerAveragingComboBox();
+    void fillMaximumUpAndDownhill();
+    void fillDifficultySetting();
     void saveVideoFolder(const QString &folder);
     bool ableToWriteInFolder(const QString &folder);