+++ mainlib/ride/actuators.cpp
@@ -1,5 +1,6 @@
 
 #include "actuators.h"
+#include "config/bigringsettings.h"
 #include "model/cyclist.h"
 
 #include "ant/antcentraldispatch.h"
@@ -7,7 +8,10 @@ namespace indoorcycling {
 
 Actuators::Actuators(const Cyclist *cyclist, indoorcycling::AntCentralDispatch *antCentralDispatch,
                                     const NamedSensorConfigurationGroup &sensorConfigurationGroup, QObject *parent):
-    QObject(parent), _cyclist(cyclist), _antCentralDispatch(antCentralDispatch), _sensorConfigurationGroup(sensorConfigurationGroup)
+    QObject(parent), _cyclist(cyclist), _antCentralDispatch(antCentralDispatch), _sensorConfigurationGroup(sensorConfigurationGroup),
+    _maximumSlope(BigRingSettings().maximumUphillForSmartTrainer()),
+    _minimumSlope(-BigRingSettings().maximumDownhillForSmartTrainer()),
+    _difficultySetting(BigRingSettings().difficultySetting() * 0.01)
 {
     connect(_antCentralDispatch, &AntCentralDispatch::sensorFound, this, &Actuators::setSensorFound);
 }
@@ -24,11 +28,15 @@ void Actuators::setSensorFound(AntSensorType channelType, int)
     }
 }
 
-void Actuators::setSlope(qreal slopeInPercent)
+void Actuators::setSlope(const qreal slopeInPercent)
 {
     if (!qFuzzyCompare(slopeInPercent, _currentSlope)) {
         _currentSlope = slopeInPercent;
-        _antCentralDispatch->setSlope(slopeInPercent);
+
+        const qreal slopeWithDifficultySetting = slopeInPercent * _difficultySetting;
+        const qreal cappedSlope = qBound(_minimumSlope, slopeWithDifficultySetting, _maximumSlope);
+
+        _antCentralDispatch->setSlope(cappedSlope);
     }
 }