+++ mainlib/video/framecopyingvideoreader.h
@@ -1,6 +1,8 @@
 #ifndef VIDEOREADER_H
 #define VIDEOREADER_H
 
+#include <memory>
+
 #include <QtCore/QEvent>
 #include <QtCore/QObject>
 #include "genericvideoreader.h"
@@ -20,7 +22,7 @@ public:
     virtual ~FrameCopyingVideoReader();
 
     void openVideoFile(const QString &videoFilename);
-    void copyNextFrame(const FrameBuffer& buffer, int skipFrames = 0);
+    void copyNextFrame(const std::weak_ptr<FrameBuffer> &buffer, int skipFrames = 0);
     void seekToFrame(qint64 frameNumber);
 
 signals:
@@ -33,7 +35,7 @@ protected:
     virtual bool event(QEvent *);
 private:
     virtual void openVideoFileInternal(const QString &videoFilename);
-    void copyNextFrameInternal(const FrameBuffer &buffer, int skipFrames);
+    void copyNextFrameInternal(std::weak_ptr<FrameBuffer> &buffer, int skipFrames);
     void seekToFrameInternal(const qint64 frameNumber);
     QSize totalFrameSize();